[WIP] Unreal Logger
If you have numerous builds of the same game like myself, your official Unreal logs get stored in the same spot, overlapping and overwriting each other.

The plan is to retrieve the point where the logs are sent from within the direct instance that UE4 Librarian is attached to,  and intercepting those logs without modifying them to copy them into organized logs located within the same directory as UE4 Librarian.

And you might ask, why not just copy the original log file? If you have multiple instances of a game only one of them will be written, unless you have a game in which the developers implemented multiple logs, something that is lacking from most games for some reason.

And for reverse engineering, logs, are ESSENTIAL.

If I manage to get that far, I plan to also figure out a way to force the game to output VeryVerbose logs, which provide the most detail in Unreal.
You can force the game to get VeryVerbose logs by modding it using a pak bypass, however this may not always be an option.